% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable_helpers.R
\name{dui_add_reactable_dep}
\alias{dui_add_reactable_dep}
\title{Add 'dataui' Dependency to 'reactable'}
\usage{
dui_add_reactable_dep(rt)
}
\arguments{
\item{rt}{\code{reactable} htmlwidget}
}
\value{
\code{reactable} htmlwidget with 'dataui' dependency attached
}
\description{
Add 'dataui' Dependency to 'reactable'
}
\examples{
if(requireNamespace("reactable")) {
   dui_add_reactable_dep(reactable(mtcars))
}
}
