% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{dui_densityseries}
\alias{dui_densityseries}
\title{Density Line Series for Histograms}
\usage{
dui_densityseries(rawData = NULL, ...)
}
\arguments{
\item{rawData}{\code{vector}, \code{list}, \code{data.frame} of data}

\item{...}{see \href{https://github.com/williaster/data-ui/tree/master/packages/histogram#series-}{data-ui series}}
}
\value{
\code{reactR component} which is a special form of \code{htmltools::tag}
}
\description{
Density Line Series for Histograms
}
