% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable_helpers.R
\name{dui_for_reactable}
\alias{dui_for_reactable}
\title{Convert 'dataui' to 'reactable' Custom Render Function}
\usage{
dui_for_reactable(dui, jsarg = "cellInfo")
}
\arguments{
\item{dui}{\code{dataui} htmlwidget to convert}
}
\value{
\code{htmlwidgets::JS} of class \code{JS_EVAL}
}
\description{
Convert 'dataui' to 'reactable' Custom Render Function
}
\examples{
if(requireNamespace("reactable")) {
  # create data frame with a list column
  df <- data.frame(x=1)
  df$x[1] <- list(1:10)
  dui_add_reactable_dep(
    reactable::reactable(
      df,
      columns = list(
        x = reactable::colDef(
          cell = dui_for_reactable(dui_sparkline(
            data = htmlwidgets::JS("cellInfo.value"),
            components=list( dui_sparklineseries() )
          )
        ))
      )
    )
  )
}
}
