% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duihistogram.R
\name{dui_histogram}
\alias{dui_histogram}
\title{<Add Title>}
\usage{
dui_histogram(
  rawData = NULL,
  binCount = NULL,
  binType = NULL,
  binValues = NULL,
  cumulative = NULL,
  horizontal = NULL,
  limits = NULL,
  margin = NULL,
  normalized = NULL,
  renderTooltip = NULL,
  valueAccessor = NULL,
  onMouseMove = NULL,
  onMouseLeave = NULL,
  tooltipData = NULL,
  ariaLabel = NULL,
  components = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{elementId}{}
}
\description{
<Add Description>
}
