% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{dui_sparkbandline}
\alias{dui_sparkbandline}
\title{Band Lines for 'Sparklines'}
\usage{
dui_sparkbandline(...)
}
\arguments{
\item{...}{see \href{https://github.com/williaster/data-ui/tree/master/packages/sparkline#reference-lines-and-bands}{data-ui reference lines and bands}}
}
\value{
\code{reactR component} which is a special form of \code{htmltools::tag}
}
\description{
Band Lines for 'Sparklines'
}
