% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{dui_sparkbarseries}
\alias{dui_sparkbarseries}
\title{Bar Series for 'Sparklines'}
\usage{
dui_sparkbarseries(...)
}
\arguments{
\item{...}{see \href{https://github.com/williaster/data-ui/tree/master/packages/sparkline#series}{data-ui series}}
}
\value{
\code{reactR component} which is a special form of \code{htmltools::tag}
}
\description{
Bar Series for 'Sparklines'
}
