% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{dui_sparklabel}
\alias{dui_sparklabel}
\title{Labels for 'Sparklines'}
\usage{
dui_sparklabel(...)
}
\arguments{
\item{...}{see \href{https://github.com/williaster/data-ui/blob/master/packages/sparkline/src/annotation/Label.jsx}{data-ui label}}
}
\value{
\code{reactR component} which is a special form of \code{htmltools::tag}
}
\description{
Labels for 'Sparklines'
}
