% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duisparkline.R
\name{dui_sparkline_with_tooltip}
\alias{dui_sparkline_with_tooltip}
\title{Title}
\usage{
dui_sparkline_with_tooltip(
  data = NULL,
  className = NULL,
  margin = NULL,
  max = NULL,
  min = NULL,
  onMouseMove = NULL,
  onMouseLeave = NULL,
  preserveAspectRatio = NULL,
  valueAccessor = NULL,
  viewBox = NULL,
  ariaLabel = NULL,
  renderTooltip = NULL,
  components = list(),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{elementId}{}
}
\value{

}
\description{
Title
}
