% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{dui_sparklineargradient}
\alias{dui_sparklineargradient}
\title{Linear Gradient for 'Sparklines'}
\usage{
dui_sparklineargradient(...)
}
\arguments{
\item{...}{see \href{https://github.com/williaster/data-ui/tree/master/packages/sparkline#patternlines--and-lineargradients}{data-ui patterns and gradients}}
}
\value{
\code{reactR component} which is a special form of \code{htmltools::tag}
}
\description{
Linear Gradient for 'Sparklines'
}
