% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{dui_sparklineseries}
\alias{dui_sparklineseries}
\title{Line Series for 'Sparklines'}
\usage{
dui_sparklineseries(...)
}
\arguments{
\item{...}{see \href{https://github.com/williaster/data-ui/tree/master/packages/sparkline#series}{data-ui series}}
}
\value{
\code{reactR component} which is a special form of \code{htmltools::tag}
}
\description{
Line Series for 'Sparklines'
}
