% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{dui_sparkpatternlines}
\alias{dui_sparkpatternlines}
\title{Pattern Fill for 'Sparklines'}
\usage{
dui_sparkpatternlines(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Pattern Fill for 'Sparklines'
}
