% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{dui_sparkpointseries}
\alias{dui_sparkpointseries}
\title{Point Series for 'Sparklines'}
\usage{
dui_sparkpointseries(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Point Series for 'Sparklines'
}
