% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{dui_sparkrefline}
\alias{dui_sparkrefline}
\title{Horizontal Reference Line for 'Sparklines'}
\usage{
dui_sparkrefline(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Horizontal Reference Line for 'Sparklines'
}
