% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{dui_tooltip}
\alias{dui_tooltip}
\title{Tooltip Container for 'Sparklines'}
\usage{
dui_tooltip(components)
}
\arguments{
\item{components}{}
}
\value{

}
\description{
Tooltip Container for 'Sparklines'
}
