% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{dui_xaxis}
\alias{dui_xaxis}
\title{'X' Axis for Histograms}
\usage{
dui_xaxis(...)
}
\arguments{
\item{...}{see \href{https://github.com/williaster/data-ui/tree/master/packages/histogram#xaxis--and-yaxis-}{data-ui axis}}
}
\value{
\code{reactR component} which is a special form of \code{htmltools::tag}
}
\description{
'X' Axis for Histograms
}
