% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hist_to_binned_data}
\alias{hist_to_binned_data}
\title{Convert Histogram to Binned Format}
\usage{
hist_to_binned_data(h = NULL, density = TRUE)
}
\arguments{
\item{h}{histogram object or list from \code{\link[graphics]{hist}}}

\item{density}{\code{logical} to indicate the use of \code{density} or \code{count}}
}
\value{
\code{list}
}
\description{
This is a convenience function to convert a histogram object from \code{\link[graphics]{hist}}
into the binned value format expected by 'data-ui' \code{\link{dui_barseries}}
and \code{\link{dui_densityseries}} argument \code{binnedData}.
}
\examples{
library(dataui)

h <- graphics::hist(stats::rnorm(100), plot = FALSE)
dui_histogram(
  components = list(
    dui_barseries(binnedData = hist_to_binned_data(h))
  )
)
}
