% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency.R
\name{html_dependency_dataui}
\alias{html_dependency_dataui}
\title{'JavaScript' Dependencies for Standalone Usage}
\usage{
html_dependency_dataui()
}
\value{

}
\description{
'JavaScript' Dependencies for Standalone Usage
}
