% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duihistogram.R
\name{dataui-shiny}
\alias{dataui-shiny}
\alias{datauiOutput}
\alias{renderDataui}
\alias{dataui_html}
\title{Shiny bindings for dataui}
\usage{
datauiOutput(outputId, width = "100\%", height = "400px")

renderDataui(expr, env = parent.frame(), quoted = FALSE)

dataui_html(id, style, class, ...)
}
\arguments{
\item{outputId}{output variable to read from}

\item{expr}{An expression that generates a dataui}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using dataui within Shiny
applications and interactive Rmd documents.
}
