% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{dui_densityseries}
\alias{dui_densityseries}
\title{Density Line Series for Histograms}
\usage{
dui_densityseries(rawData = NULL, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Density Line Series for Histograms
}
