% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duisparkline.R
\name{dui_sparkline}
\alias{dui_sparkline}
\title{'data-ui Sparklines'}
\usage{
dui_sparkline(
  data = NULL,
  className = NULL,
  margin = NULL,
  max = NULL,
  min = NULL,
  onMouseMove = NULL,
  onMouseLeave = NULL,
  renderTooltip = NULL,
  preserveAspectRatio = NULL,
  valueAccessor = NULL,
  viewBox = NULL,
  ariaLabel = NULL,
  components = list(),
  width = 300,
  height = 100,
  elementId = NULL
)
}
\arguments{
\item{elementId}{}
}
\value{
\code{htmlwidget}
}
\description{
Create interactive 'sparklines' visualizations for use as a standalone comoponent or
in combination with 'reactable' or other table libraries.  There are a wide variety of
components for use with \code{dui_sparkline}, including lines, bars, points, reference
lines, and labels.
}
