% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duisparkline.R
\name{dui_sparkline}
\alias{dui_sparkline}
\title{'data-ui Sparklines'}
\usage{
dui_sparkline(
  data = NULL,
  className = NULL,
  margin = NULL,
  max = NULL,
  min = NULL,
  onMouseMove = NULL,
  onMouseLeave = NULL,
  renderTooltip = NULL,
  preserveAspectRatio = NULL,
  valueAccessor = NULL,
  viewBox = NULL,
  ariaLabel = NULL,
  components = list(),
  responsive = TRUE,
  width = "100\%",
  height = 100,
  elementId = NULL
)
}
\arguments{
\item{responsive}{\code{logical} with default as \code{TRUE}.  This argument is not passed
on to the \code{React} component as a \code{prop}.  Rather it is used in the \code{R}
constructor to choose our \code{SparklineResponsive} component (\code{TRUE}) or
\code{SparklineWithTooltip} (\code{FALSE}) component.  Since \code{height} can be unstable
and finicky, we only make \code{width} responsive.}

\item{elementId}{}
}
\value{
\code{htmlwidget}
}
\description{
Create interactive 'sparklines' visualizations for use as a standalone comoponent or
in combination with 'reactable' or other table libraries.  There are a wide variety of
components for use with \code{dui_sparkline}, including lines, bars, points, reference
lines, and labels.  These 'sparklines' automatically include responsive and tooltip
behaviors.
}
