% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duisparkline.R
\name{dui_sparkline}
\alias{dui_sparkline}
\title{<Add Title>}
\usage{
dui_sparkline(
  data = NULL,
  className = NULL,
  margin = NULL,
  max = NULL,
  min = NULL,
  onMouseMove = NULL,
  onMouseLeave = NULL,
  preserveAspectRatio = NULL,
  valueAccessor = NULL,
  viewBox = NULL,
  ariaLabel = NULL,
  components = list(),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{elementId}{}
}
\description{
<Add Description>
}
