% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{dui_sparklineseries}
\alias{dui_sparklineseries}
\title{Line Series for 'Sparklines'}
\usage{
dui_sparklineseries(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Line Series for 'Sparklines'
}
