% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{dui_sparkverticalrefline}
\alias{dui_sparkverticalrefline}
\title{Vertical Reference Line for 'Sparklines'}
\usage{
dui_sparkverticalrefline(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Vertical Reference Line for 'Sparklines'
}
