% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{dui_tooltip}
\alias{dui_tooltip}
\title{Tooltip Container for 'Sparklines'}
\usage{
dui_tooltip(components)
}
\arguments{
\item{components}{\code{list} of children (series or reference lines) to include
in the sparkline.  Multiple components should be wrapped in \code{list} such as
\code{components = list(dui_sparklineseries, dui_sparkverticalrefline())}.}
}
\value{
\code{reactR component} which is a special form of \code{htmltools::tag}
}
\description{
Tooltip Container for 'Sparklines'
}
